Add the library curl
To satisfy undefined references coming from libnetcdf.

This web page (https://www.unidata.ucar.edu/software/netcdf/docs/getting_and_building_netcdf.html)
indicates that the curl library is 'for DAP remote access client support'.  Because we do not
deal with remote clients, the curl library is of no consequence for us other than getting the
r.diag code to compile.  Therefore, although the curl library that is linked might not be correct
for our compiler, it does not matter.