Fix post-install scripts
We need to load Intel compiler for rhel-9-graniterapids-64, as some libraries necessary for netcdf are not found.
Moreover, we need to load udunits2 from rpn domain and export DUNITS2_XML_PATH to the proper path.

Fix some errors in the csh script, but it surely is never used, as the r.diag_autocomplete.dot sourced in the scripts is not csh compatible